- dao parameter type 수정

+++ src/main/java/com/aispace/erksystem/service/database/ServiceProviderDAO.java
@@ -82,7 +82,7 @@ public class ServiceProviderDAO {
         }
     }
 
-    public static boolean delete(long orgId) {
+    public static boolean delete(int orgId) {
         Session session = sessionFactory.openSession();
         Transaction tx = null;
         try {
